google-cloud-ai-certifications/professional-ml-engineer/code-examples/deployment-scripts/fastapi-vertex-endpoint/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-cloud-aiplatform>=1.71.1
